[34-navigation-desktop]: Navbar folder sepration Separated out desktop navigation files and mobile navigation files in separate folders.
